requirements/develop.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file develop.txt develop.in
#
click==6.7                # via pip-tools
cookies==2.2.1            # via responses
first==2.0.1              # via pip-tools
future==0.16.0
oauthlib==1.0.3
pip-tools==1.9.0
py==1.4.33                # via pytest
pytest==3.0.7
python-twitter==3.2.1
requests-oauthlib==0.6.0
requests==2.9.0
responses==0.5.1
six==1.10.0               # via pip-tools, responses
